Back development

+++ src/controllers/bedsController.ts
@@ -1,6 +1,6 @@
 import { PrismaClient } from "@prisma/client";
 import { v4 as uuidv4 } from "uuid";
-import { bedNameValidator, bedSectionValidator } from "../error/bedsErrorHandler";
+import { bedNameValidator, bedSectionValidator, checkPatient } from "../error/bedsErrorHandler";
 
 const prisma = new PrismaClient();
 
@@ -175,11 +175,20 @@ export async function updateBed(req, res){
 export async function deleteBed(req, res){
   const { bedId } = req.body
 
+  if (await checkPatient(bedId)){
+    res.status = 400
+    res.send({
+      Error: "Leito Ocuppaded"
+    })
+    return
+  }
+
   const deleteBedHistoric = await prisma.bedHistoric.deleteMany({
     where:{
       bedsId: bedId
     }
   })
+  
 
   const deleteBed = await prisma.beds.delete({
     where:{
